tests/wpt/tests/workers/postMessage_DataCloneErr.htm

<!DOCTYPE html>
<title> postMessage() with WorkerNavigator </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function(t) {
  var worker = new Worker("./support/WorkerDataCloneErr.js");
  worker.onmessage = t.step_func_done(function(e) {
    assert_equals(e.data, "Pass");
  });
});
</script>
